rust: Move tests indoors
