+++ rust/foxglove/src/lib.rs
@@ -154,6 +154,8 @@
 //!
 //! [tokio]: https://docs.rs/tokio/latest/tokio/
 
+//#![warn(missing_docs)]
+
 use thiserror::Error;
 
 mod channel;
@@ -164,40 +166,52 @@ mod encode;
 mod log_context;
 mod log_sink;
 mod log_sink_set;
-#[cfg(test)]
-mod log_sink_test;
 mod mcap_writer;
 mod metadata;
 pub mod schemas;
 mod time;
-#[doc(hidden)]
-pub mod websocket;
+mod websocket;
 mod websocket_server;
+
+#[cfg(test)]
+mod tests;
+#[cfg(test)]
+mod testutil;
 
 pub use channel::{Channel, Schema};
 pub use channel_builder::ChannelBuilder;
 pub use encode::{Encode, TypedChannel};
 // For tests
 #[doc(hidden)]
-pub use log_context::{GlobalContextTest, LogContext};
+pub use log_context::LogContext;
 pub use log_sink::LogSink;
 pub use mcap_writer::{McapWriter, McapWriterHandle};
 pub use metadata::{Metadata, PartialMetadata};
 pub(crate) use time::nanoseconds_since_epoch;
+#[doc(hidden)]
+#[cfg(feature = "unstable")]
+pub use websocket::{Parameter, ParameterType, ParameterValue};
 pub use websocket_server::{WebSocketServer, WebSocketServerHandle};
 
+/// An error type for errors generated by this crate.
 #[derive(Error, Debug)]
 #[non_exhaustive]
 pub enum FoxgloveError {
+    /// A unspecified unrecoverable error.
     #[error("Fatal error: {0}")]
     Fatal(String),
+    /// A channel for the same topic has already been registered.
     #[error("Channel for topic {0} already exists in registry")]
     DuplicateChannel(String),
+    /// An I/O error.
     #[error(transparent)]
     IOError(#[from] std::io::Error),
+    /// An error related to MCAP encoding.
     #[error("MCAP error: {0}")]
     MCAPError(#[from] mcap::McapError),
-    #[allow(dead_code)]
+    /// An error related to JSON encoding.
+    #[doc(hidden)]
+    #[cfg(feature = "unstable")]
     #[error(transparent)]
     JSONError(#[from] serde_json::Error),
 }
