+++ rust/foxglove/src/tests.rs
@@ -1,0 +1,4 @@
+mod logging;
+mod websocket;
+#[cfg(feature = "unstable")]
+mod websocket_unstable;
