+++ rust/foxglove/src/testutil.rs
@@ -1,0 +1,7 @@
+//! Test utilities.
+
+mod log_context;
+mod log_sink;
+
+pub use log_context::GlobalContextTest;
+pub use log_sink::{ErrorSink, MockSink, RecordingSink};
