+++ rust/foxglove/src/websocket.rs
@@ -31,9 +31,7 @@ mod protocol;
 #[cfg(test)]
 mod tests;
 
-// For tests
-#[doc(hidden)]
-pub const SUBPROTOCOL: &str = "foxglove.sdk.v1";
+pub(crate) const SUBPROTOCOL: &str = "foxglove.sdk.v1";
 
 type WebsocketSender = SplitSink<WebSocketStream<TcpStream>, Message>;
 
@@ -65,34 +63,14 @@ fn get_tokio_runtime() -> Handle {
     runtime.handle().clone()
 }
 
-#[doc(hidden)]
-pub struct InternalServerOptions {
+#[derive(Default)]
+pub(crate) struct ServerOptions {
     pub session_id: Option<String>,
     pub name: Option<String>,
-    pub listener: Option<Arc<dyn ServerListener>>,
     pub message_backlog_size: Option<usize>,
+    pub listener: Option<Arc<dyn ServerListener>>,
     pub capabilities: Option<HashSet<Capability>>,
     pub supported_encodings: Option<HashSet<String>>,
-}
-
-#[derive(Default)]
-pub struct ServerOptions {
-    pub session_id: Option<String>,
-    pub name: Option<String>,
-    pub message_backlog_size: Option<usize>,
-}
-
-impl From<ServerOptions> for InternalServerOptions {
-    fn from(options: ServerOptions) -> Self {
-        Self {
-            session_id: options.session_id,
-            name: options.name,
-            message_backlog_size: options.message_backlog_size,
-            listener: None,
-            capabilities: None,
-            supported_encodings: None,
-        }
-    }
 }
 
 impl std::fmt::Debug for ServerOptions {
@@ -315,7 +293,7 @@ impl ConnectedClient {
 
 // A websocket server that implements the Foxglove WebSocket Protocol
 impl Server {
-    pub(crate) fn new(weak_self: Weak<Self>, opts: InternalServerOptions) -> Self {
+    pub(crate) fn new(weak_self: Weak<Self>, opts: ServerOptions) -> Self {
         Server {
             weak_self,
             started: AtomicBool::new(false),
@@ -750,11 +728,6 @@ impl LogSink for Server {
 
 #[doc(hidden)]
 pub fn create_server(opts: ServerOptions) -> Arc<Server> {
-    Arc::new_cyclic(|weak_self| Server::new(weak_self.clone(), opts.into()))
-}
-
-#[cfg(feature = "unstable")]
-pub fn create_server_with_internal_options(opts: InternalServerOptions) -> Arc<Server> {
     Arc::new_cyclic(|weak_self| Server::new(weak_self.clone(), opts))
 }
